fix conftest.py, update uv.lock
